SpectralFactorMOR

Documentation for SpectralFactorMOR.

API

SpectralFactorMOR.AlmostKroneckerDAE

SpectralFactorMOR.SemiExplicitIndex1DAE

SpectralFactorMOR.StaircaseDAE

SpectralFactorMOR.arec_lr_nwt

SpectralFactorMOR.checkprojlure

SpectralFactorMOR.compress_lr

SpectralFactorMOR.compress_lr

SpectralFactorMOR.initial_shifts

SpectralFactorMOR.irka

SpectralFactorMOR.irka

SpectralFactorMOR.irka

SpectralFactorMOR.is_valid

SpectralFactorMOR.isastable

SpectralFactorMOR.lrcf

SpectralFactorMOR.lyapc_lradi

SpectralFactorMOR.new_shifts

SpectralFactorMOR.pr_c_gramian

SpectralFactorMOR.pr_c_gramian

SpectralFactorMOR.pr_c_gramian_lr

SpectralFactorMOR.pr_c_gramian_lr

SpectralFactorMOR.pr_o_gramian

SpectralFactorMOR.pr_o_gramian

SpectralFactorMOR.pr_o_gramian_lr

SpectralFactorMOR.pr_o_gramian_lr

SpectralFactorMOR.prbaltrunc

SpectralFactorMOR.retry_irka

SpectralFactorMOR.sfmor

SpectralFactorMOR.sfmor

SpectralFactorMOR.splitsys

SpectralFactorMOR.splitsys

SpectralFactorMOR.splitsys

SpectralFactorMOR.toindex0

SpectralFactorMOR.tokronecker

SpectralFactorMOR.truncation

SpectralFactorMOR.AlmostKroneckerDAE — Type

            |E_11    0  0 0|      | 0   0  0 I|      |B_1|
        E = |  0   E_22 0 0|, A = | 0 A_22 0 0|, B = |B_2|, C = |C_1 C_2 C_3 C_4|
            |  0     0  0 0|      | 0   0  I 0|      |B_3|
            |  0     0  0 0|      |-I   0  0 0|      |B_4|

The matrices E11 and E22 are nonsingular.

SpectralFactorMOR.SemiExplicitIndex1DAE — Type

        | E_11 0 | 
    E = |        | 
        | 0    0 |

        | A_11 A_12 |
    A = |           |
        | A_21 A_22 |

        | B_1 |
    B = |     |
        | B_2 |

    C = | C_1 C_2 |

The matrix E_11 is nonsingular.

SpectralFactorMOR.StaircaseDAE — Type

            |E_11    0  0 0|      |A_11 A_12 A_13 A_14|      |B_1|
        E = |  0   E_22 0 0|, A = |A_21 A_22 A_23   0 |, B = |B_2|, C = |C_1 C_2 C_3 C_4|
            |  0     0  0 0|      |A_31 A_32 A_33   0 |      |B_3|
            |  0     0  0 0|      |A_41   0    0    0 |      |B_4|

with E_11, E_22, A_14=-A_41^T, A_33 nonsingular

SpectralFactorMOR.arec_lr_nwt — Function

arec_lr_nwt(F, E, Q, R, P_r=I, P_l=I; 
conv_tol=1e-12, max_iterations=20, 
lyapc_lradi_tol=1e-12, lyapc_lradi_max_iterations=150)

Low-rank Newton method [DAEForumIV, Algorithm 9].

FXE^T + EXF^T + EXQ^TQXE^T + P_l RR^T P_l^T = 0,
X = P_r X P_r^T.

Returns an approximate solution Z such that

X ≈ Z*Z'.

SpectralFactorMOR.checkprojlure — Method

Check if X satisfies the projected positive real Lur'e equation.

SpectralFactorMOR.compress_lr — Method

compress_lr(Z)

Compress Z via SVD such that it holds ZZ' ≈ Z_newZ_new'.

SpectralFactorMOR.compress_lr — Method

compress_lr(Z)

Compress Z via SVD such that it holds ZZ' ≈ Z_newZ_new'.

SpectralFactorMOR.initial_shifts — Function

initial_shifts(A,E,B,max_iterations=10,iter=0)

Returns initial shifts for LR-ADI method.

See [Kür16_Doktorarbeit, pp. 92-95] and https://github.com/pymor/pymor/blob/main/src/pymor/algorithms/lradi.py.

SpectralFactorMOR.irka — Method

i0irka(
    sys::Union{DescriptorStateSpace,SparseDescriptorStateSpace},
    r::Int, irka_options::IRKAOptions
)

IRKA for index-0 systems.

SpectralFactorMOR.irka — Method

i1irka(sys::SemiExplicitIndex1DAE, r::Int, irka_options::IRKAOptions)

IRKA for semiexplicit index-1 DAEs.

Gugercin, S., Stykel, T., & Wyatt, S. (2013). Model Reduction of Descriptor Systems by Interpolatory Projection Methods. SIAM. https://doi.org/10.1137/130906635

SpectralFactorMOR.irka — Method

i1irka(sys::StaircasePHDAE, r::Int, irka_options::IRKAOptions)

IRKA for port-Hamiltonian DAEs in staircase form.

SpectralFactorMOR.is_valid — Method

Assert that sys is in valid staircase form.

SpectralFactorMOR.isastable — Method

Check if sys is asymptotically stable.

SpectralFactorMOR.lrcf — Method

lrcf(X, trunc_tol)

Computes an approximate low-rank Cholesky-like factorization of a symmetric positive semi-definite matrix X s.t. X = Z'*Z (up to a prescribed tolerance trunc_tol).

SpectralFactorMOR.lyapc_lradi — Function

lyapc_lradi(A, E, B, P_l=I; tol=1e-12, max_iterations=150)

LR-ADI method for the projected continuous-time Lyapunov equation [DAEForumIV, Algorithm 7]

EXA^T + AXE^T = - P_l BB^T P_l^T,
X = P_r X P_r^T.

Returns an approximate solution Z such that

X ≈ Z*Z'.

SpectralFactorMOR.new_shifts — Function

new_shifts(A,E,V,Z,prev_shifts,subspace_columns=6)

Returns new intermediate shifts for LR-ADI method.

See [Kür16_Doktorarbeit, pp. 92-95] and https://github.com/pymor/pymor/blob/main/src/pymor/algorithms/lradi.py.

SpectralFactorMOR.pr_c_gramian — Method

pr_c_gramian(sys)

Computes the positive real observability gramian, i.e., the unique stabilizing solution of

AXE^T + EXA^T + (P_l B - EXC^T) (M_0+M_0')^{-1} (P_l B - EXC^T)^T = 0,
X = P_r X P_r^T,

where Pr, Pl and M_0 are defined by the given system.

Note: Uses MatrixEquations.garec.

SpectralFactorMOR.pr_c_gramian — Method

pr_c_gramian(sys)

Computes the positive real observability gramian, i.e., the unique stabilizing solution of

AXE^T + EXA^T + (P_l B - EXC^T) (M_0+M_0')^{-1} (P_l B - EXC^T)^T = 0,
X = P_r X P_r^T,

where Pr, Pl and M_0 are defined by the given system.

Note: Uses MatrixEquations.garec.

SpectralFactorMOR.pr_c_gramian_lr — Method

pr_c_gramian_lr(sys)

Computes low-rank factor of positive real controllability gramian, i.e., the unique stabilizing solution of

AXE^T + EXA^T + (B - EXC^T) (D+D')^{-1} (B - EXC^T)^T = 0

E is assumed to be nonsingular. Returns an approximate solution Z such that

X ≈ Z*Z'.

SpectralFactorMOR.pr_c_gramian_lr — Method

pr_c_gramian_lr(sys)

Computes low-rank factor of positive real observability gramian, i.e., the unique stabilizing solution of

AXE^T + EXA^T + (P_l B - EXC^T) (M_0+M_0')^{-1} (P_l B - EXC^T)^T = 0,
X = P_r X P_r^T,

where Pr, Pl and M_0 are defined by the given system.

Returns an approximate solution Z such that

X ≈ Z*Z'.

SpectralFactorMOR.pr_o_gramian — Method

pr_o_gramian(sys)

Computes positive real observability gramian, i.e., the unique stabilizing solution of

A^TXE + E^TXA + (P_r^T C^T-E^TXB)(M_0+M_0')^{-1}(C P_r-B^TXE) = 0,
X = P_l^T Y P_l,

where Pr, Pl and M_0 are defined by the given system.

Note: Uses MatrixEquations.garec.

SpectralFactorMOR.pr_o_gramian — Method

pr_o_gramian(sys)

Computes positive real observability gramian, i.e., the unique stabilizing solution of

A^TXE + E^TXA + (P_r^T C^T-E^TXB)(M_0+M_0')^{-1}(C P_r-B^TXE) = 0,
X = P_l^T Y P_l,

where Pr, Pl and M_0 are defined by the given system.

Note: Uses MatrixEquations.garec.

SpectralFactorMOR.pr_o_gramian_lr — Method

pr_o_gramian_lr(sys)

Computes low-rank factor of positive real observability gramian, i.e., the unique stabilizing solution of

A^TXE + E^TXA + (C^T-E^TXB)(D+D')^{-1}(C-B^TXE) = 0.

E is assumed to be nonsingular. Returns an approximate solution Z such that

X ≈ Z*Z'.

SpectralFactorMOR.pr_o_gramian_lr — Method

pr_o_gramian_lr(sys)

Computes low-rank factor of positive real observability gramian, i.e., the unique stabilizing solution of

A^TXE + E^TXA + (P_r^T C^T-E^TXB)(M_0+M_0')^{-1}(C P_r-B^TXE) = 0,
X = P_l^T Y P_l,

where Pr, Pl and M_0 are defined by the given system.

Returns an approximate solution Z such that

X ≈ Z*Z'.

SpectralFactorMOR.prbaltrunc — Method

prbaltrunc(sys::SemiExplicitIndex1DAE, r::Int)

Positive real balanced truncation for descriptor systems [DAEForumIV, Algorithm 2].

The parameter r corresponds to the reduced order of the finite part of the system.

SpectralFactorMOR.retry_irka — Method

Retry IRKA max_retries times and pick the result with the lowest H2-error. syssp is the strictly proper subsystem of the full-order model.

SpectralFactorMOR.sfmor — Method

A solution to the positive real projected Lur'e equation can be supplied via the parameter X. If no X is supplied, the positive real observability Gramian is computed.

SpectralFactorMOR.sfmor — Method

A solution to the positive real projected Lur'e equation can be supplied via the parameter X. If no X is supplied, the positive real observability Gramian is computed.

SpectralFactorMOR.splitsys — Method

splitsys(sys::SemiExplicitIndex1DAE)

Return infinite and strictly proper subsystems.

sys1 is the infinite subsystem.

sys2 is the strictly proper subsystem.

SpectralFactorMOR.splitsys — Method

splitsys(sys::AlmostKroneckerDAE)

Return infinite and strictly proper subsystems.

sys1 is the infinite subsystem.

sys2 is the strictly proper subsystem.

SpectralFactorMOR.splitsys — Method

splitsys(sys::StaircaseDAE)

Return infinite and strictly proper subsystems.

sys1 is the infinite subsystem.

sys2 is the strictly proper subsystem.

SpectralFactorMOR.toindex0 — Method

toindex0(sys::SemiExplicitIndex1DAE)

Transform sys to a system with index 0.

SpectralFactorMOR.tokronecker — Method

Transform sys into almost Kronecker form [AchAM2021].

SpectralFactorMOR.truncation — Method

truncation(d, L, trunc_tol)

Computes a rank revealing factorization for a given LDL-decomposition of $S = L * diagm(d) * L'$ (up to a prescribed tolerance trunc_tol).

Returns dr,Lr such that $Lr * diagm(d) * Lr' \approx S$.

References